requirements.txt
langchain 
langchain-openai
langsmith 
pandas 
langchain_experimental 
matplotlib 
langgraph 
langchain_core
yfinance
streamlit
python-dotenv
langchain
openai>=1.0.0
slack-sdk 
slack-bolt 
flask
PyQt5
yahoofinance
seaborn
google-cloud-aiplatform 
langchain-google-vertexai
pydantic==2.9.0
gcloud
google-cloud-iam
playwright 
openpyxl
spacy
requests
tqdm
numpy
openai>=1.0.0
numpy>=1.21.0
spacy>=3.0.0
langchain>=0.1.0
langgraph>=0.0.10
python-dotenv>=0.19.0 
langchain_ollama
playwright
fastmcp>=2.6.0
transformers>=4.30.0
torch>=2.0.0
httpx>=0.25.0
pytest>=7.0.0
pytest-asyncio>=0.21.0